docsense/requirements.txt
# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# Embeddings (ONNX-based, no PyTorch)
fastembed>=0.3.1

# Vector DB client (connects to standalone binary)
qdrant-client>=1.9.0

# Document parsing
pymupdf>=1.24.0        # PDF
python-docx>=1.1.0     # DOCX
openpyxl>=3.1.2        # XLSX
python-pptx>=0.6.23    # PPTX

# File watching
watchdog>=4.0.0

# HTTP (for Qdrant binary download in start.py)
httpx>=0.27.0
